requirements.txt
fastapi
uvicorn
python-dotenv
httpx
python-multipart 